requirements.txt
airtable==0.3.1
alabaster==0.7.10
amqp==2.2.2
appnope==0.1.2
asn1crypto==0.22.0
attrs==21.4.0
Babel==2.4.0
backcall==0.2.0
backports.csv==1.0.7
bcrypt==3.1.3
billiard==3.5.0.3
boto==2.48.0
boto3==1.7.55
botocore==1.10.55
cassandra-driver==3.19.0
celery==4.1.0
certifi==2017.4.17
cffi==1.10.0
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.1.4
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20200713
Django==1.11.3
django-admin-rangefilter==0.3.7
django-cache-memoize==0.1.6
django-cors-headers==3.2.0
django-debug-toolbar==1.8
django-dynamic-db-router @ git+https://github.com/voltlines/django-dynamic-db-router.git@6496f7d27c5a6d5c67eea6c29581f969b35df123
django-extensions==1.8.0
django-filter==1.1.0
django-import-export==1.2.0
django-monthfield==0.1.1
django-multiselectfield==0.1.8
django-redis==4.8.0
django-rest-swagger==2.2.0
django-ses==0.8.5
django-storages==1.7.1
django-suit==0.2.25
django-table-logger @ git+https://github.com/voltlines/django-table-logger.git@ba1b059e080fe5fd3a8a46312ef2fcb78fadb375
django-user-agents==0.3.2
djangorestframework==3.6.3
docutils==0.13.1
elasticsearch==6.3.1
et-xmlfile==1.1.0
Fabric3==1.13.1.post1
facebook-sdk==2.0.0
factory-boy==2.11.1
Faker==11.3.0
flower==0.9.2
fluent-logger==0.9.3
future==0.18.2
geopy==1.11.0
gunicorn==19.7.1
helper==2.5.0
idna==2.5
imagesize==0.7.1
importlib-metadata==4.8.3
iniconfig==1.1.1
ipdb==0.13.9
ipython==7.16.2
ipython-genutils==0.2.0
itypes==1.2.0
jedi==0.17.2
Jinja2==2.9.6
jmespath==0.10.0
kinesis-python==0.1.7
kombu==4.1.0
MarkupSafe==2.0.0
mock==4.0.3
month==1.2.0
msgpack==1.0.3
ndg-httpsclient==0.5.1
newrelic==3.2.1.93
newrelic-plugin-agent==1.3.0
numpy==1.19.5
odfpy==1.4.1
offspring==0.1.1
openapi-codec==1.3.2
openpyxl==3.0.9
packaging==16.8
pandas==0.24.2
paramiko==2.2.1
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==5.0.0
pluggy==1.0.0
prompt-toolkit==2.0.10
psycopg2==2.7.1
ptyprocess==0.7.0
pusher==2.0.0
py==1.11.0
pyasn1==0.2.3
pycparser==2.18
Pygments==2.2.0
PyNaCl==1.1.2
pyOpenSSL==17.5.0
pyparsing==3.0.6
pytest==6.2.5
python-dateutil==2.8.2
pytz==2017.2
PyYAML==6.0
raven==6.8.0
redis==2.10.5
requests==2.18.1
s3transfer==0.1.13
selenium==3.8.1
Shapely==1.8.0
simplejson==3.17.6
six==1.16.0
snowballstemmer==1.2.1
Sphinx==1.6.3
sphinxcontrib-websupport==1.0.1
sqlparse==0.2.3
tablechangelogger==0.0.7
tablib==0.13.0
text-unidecode==1.3
toml==0.10.2
tornado==4.5.2
traitlets==4.3.3
typing==3.6.1
typing_extensions==4.0.1
ua-parser==0.10.0
Unidecode==1.0.22
uritemplate==4.1.1
urllib3==1.21.1
user-agents==2.2.0
validators==0.20.0
vine==1.1.4
volf @ git+https://bitbucket.org/thevoltlines/volf.git@5657175c53a074d3dad0e55110a7f86086e73079#subdirectory=volf
wcwidth==0.2.5
xlrd==1.1.0
xlwt==1.3.0
zipp==3.6.0
